requirements.txt
python-telegram-bot==20.0
requests==2.31.0
aiohttp~=3.8.5
aiogram==3.1.1
python-dotenv
fastapi
uvicorn
sqlalchemy
asyncpg
